Extract base class out of subclasses.
* Remove the rampant duplication in the various registry classes.
  Extract a base class out of them.
* Change the name from 'Collection' to 'Registry'. The new subclasses do
  not fire events, and do not contain any iteration methods.
